Dependency untangle
