+++ test/system/test_run_comparison_pipeline.py
@@ -2,9 +2,9 @@ import os
 from os.path import dirname
 from test import get_resource_dir, test_dir
 
-from ugvc.comparison.concordance_utils import read_hdf
 from ugvc.dna.format import DEFAULT_FLOW_ORDER
 from ugvc.pipelines import run_comparison_pipeline
+from ugvc.utils.misc_utils import read_hdf
 
 
 class TestRunComparisonPipeline:
