+++ test/unit/comparison/test_vcf_pipeline_utils.py
@@ -10,8 +10,8 @@ import pandas as pd
 import pysam
 from simppl.simple_pipeline import SimplePipeline
 
-from ugvc.comparison.concordance_utils import read_hdf
 from ugvc.comparison.vcf_pipeline_utils import VcfPipelineUtils, _fix_errors, bed_file_length, vcf2concordance
+from ugvc.utils.misc_utils import read_hdf
 from ugvc.vcfbed import vcftools
 from ugvc.vcfbed.interval_file import IntervalFile
