+++ ugvc/pipelines/sec/assess_sec_concordance.py
@@ -28,14 +28,10 @@ from pandas import DataFrame, Series
 from simppl.cli import get_parser
 
 from ugvc import logger
-from ugvc.comparison.concordance_utils import (
-    apply_filter,
-    calc_accuracy_metrics,
-    read_hdf,
-    validate_preprocess_concordance,
-)
+from ugvc.comparison.concordance_utils import apply_filter, calc_accuracy_metrics, validate_preprocess_concordance
 from ugvc.comparison.vcf_pipeline_utils import annotate_concordance
 from ugvc.dna.format import DEFAULT_FLOW_ORDER
+from ugvc.utils.misc_utils import read_hdf
 from ugvc.utils.stats_utils import get_precision, get_recall
